requirements.txt
## The following requirements were added by pip freeze:
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
Flask==1.1.2
Flask-Bcrypt==0.7.1
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
schema==0.7.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
